cynthia-resonance/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dateutil==2.9.0.post0
httpx==0.27.0
